+++ plot_spec.py
@@ -12,7 +12,7 @@ import matplotlib.pyplot as plt
 
 __VERSION__ = "1.0"
 
-def main(specfile,sed_plot='cont_sed.pdf',
+def main(specfile,sed_plot='cont_sed.pdf',fluxtype='flux',
          line2cont_plot='line2cont.pdf',te_plot='te.pdf'):
     """
     Read output file from calc_te.py and plot continuum flux vs.
@@ -38,23 +38,36 @@ def main(specfile,sed_plot='cont_sed.pdf',
     is_Hnalpha = np.array([lineid[0] == 'H' for lineid in data['lineid']])
     is_stacked = ~is_Hnalpha
     is_all = np.where(['all' in lineid for lineid in data['lineid']])[0][0]
+    # plot limits
+    xmin = 4000
+    xmax = 10000
+    xfit = np.linspace(xmin,xmax,100)
     #
     # Plot continuum SED
     #
+    print("Plotting continuum SED...")
     plt.ioff()
     fig, ax = plt.subplots()
-    ax.errorbar(data['frequency'][is_Hnalpha],data['cont'][is_Hnalpha],
-                yerr=data['rms'][is_Hnalpha],fmt='o',color='k')
-    # fit line 
-    fit,cov = np.polyfit(data['frequency'][is_Hnalpha],data['cont'][is_Hnalpha],1,w=1./data['rms'][is_Hnalpha],cov=True)
-    ax.plot(data['frequency'][is_Hnalpha],np.poly1d(fit)(data['frequency'][is_Hnalpha]),'k-',zorder=10,
-            label=r'$F_{{\nu,\rm C}} = ({0:.3f}\pm{1:.3f})\nu + ({2:.1f}\pm{3:.1f})$'.format(fit[0],np.sqrt(cov[0,0]),fit[1],np.sqrt(cov[1,1])))
-    # fit power law
-    fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha]),np.log10(data['cont'][is_Hnalpha]),1,w=data['cont'][is_Hnalpha]*np.log(10.)/data['rms'][is_Hnalpha],cov=True)
-    yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
-    ax.plot(data['frequency'][is_Hnalpha],yfit(data['frequency'][is_Hnalpha]),'k--',zorder=10,
-            label=r'$F_{{\nu,\rm C}} \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
-    ax.legend(loc='upper right',fontsize=10)
+    plt.xlim(xmin,xmax)
+    isnan = np.isnan(data['cont'][is_Hnalpha])
+    ax.errorbar(data['frequency'][is_Hnalpha][~isnan],
+                data['cont'][is_Hnalpha][~isnan],
+                yerr=data['rms'][is_Hnalpha][~isnan],fmt='o',color='k')
+    if len(data['cont'][is_Hnalpha][~isnan]) > 4:
+        # fit line 
+        fit,cov = np.polyfit(data['frequency'][is_Hnalpha][~isnan],
+                             data['cont'][is_Hnalpha][~isnan],1,
+                             w=1./data['rms'][is_Hnalpha][~isnan],cov=True)
+        ax.plot(xfit,np.poly1d(fit)(xfit),'k-',zorder=10,
+                label=r'$F_{{\nu,\rm C}} = ({0:.3f}\pm{1:.3f})\nu + ({2:.1f}\pm{3:.1f})$'.format(fit[0],np.sqrt(cov[0,0]),fit[1],np.sqrt(cov[1,1])))
+        # fit power law
+        fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha][~isnan]),
+                             np.log10(data['cont'][is_Hnalpha][~isnan]),1,
+                             w=data['cont'][is_Hnalpha][~isnan]*np.log(10.)/data['rms'][is_Hnalpha][~isnan],cov=True)
+        yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
+        ax.plot(xfit,yfit(xfit),'k--',zorder=10,
+                label=r'$F_{{\nu,\rm C}} \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
+        ax.legend(loc='upper right',fontsize=10)
     # plot stacked
     xlim = ax.get_xlim()
     ax.plot(xlim,
@@ -65,26 +78,42 @@ def main(specfile,sed_plot='cont_sed.pdf',
                     color='k',alpha=0.5,edgecolor='none')
     ax.set_xlim(xlim)
     ax.set_xlabel('Frequency (MHz)')
-    ax.set_ylabel('Flux Density (mJy/beam)')
+    if fluxtype == 'flux':
+        ax.set_ylabel('Flux (mJy)')
+    else:
+        ax.set_ylabel('Flux Density (mJy/beam)')
     fig.tight_layout()
     fig.savefig(sed_plot)
     plt.close(fig)
+    print("Done!")
     #
     # Plot line-to-continuum ratio vs frequency
     #
+    print("Plotting line-to-continuum ratio...")
     fig, ax = plt.subplots()
-    ax.errorbar(data['frequency'][is_Hnalpha],data['line2cont'][is_Hnalpha],
-                yerr=data['e_line2cont'][is_Hnalpha],fmt='o',color='k')
-    # fit line 
-    fit,cov = np.polyfit(data['frequency'][is_Hnalpha],data['line2cont'][is_Hnalpha],1,w=1./data['e_line2cont'][is_Hnalpha],cov=True)
-    ax.plot(data['frequency'][is_Hnalpha],np.poly1d(fit)(data['frequency'][is_Hnalpha]),'k-',zorder=10,
-            label=r'$F_{{\nu,\rm L}}/F_{{\nu,\rm C}} = ({0:.3f}\pm{1:.3f})\times10^{{-3}}\nu + ({2:.3f}\pm{3:.3f})$'.format(fit[0]*1.e3,np.sqrt(cov[0,0])*1.e3,fit[1],np.sqrt(cov[1,1])))
-    # fit power law
-    fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha]),np.log10(data['line2cont'][is_Hnalpha]),1,w=data['line2cont'][is_Hnalpha]*np.log(10.)/data['e_line2cont'][is_Hnalpha],cov=True)
-    yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
-    ax.plot(data['frequency'][is_Hnalpha],yfit(data['frequency'][is_Hnalpha]),'k--',zorder=10,
-            label=r'$F_{{\nu,\rm L}}/F_{{\nu,\rm C}} \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
-    ax.legend(loc='upper left',fontsize=10)
+    plt.xlim(xmin,xmax)
+    isnan = np.isnan(data['line2cont'][is_Hnalpha])
+    ax.errorbar(data['frequency'][is_Hnalpha][~isnan],data['line2cont'][is_Hnalpha][~isnan],
+                yerr=data['e_line2cont'][is_Hnalpha][~isnan],fmt='o',color='k')
+    # limits
+    ax.plot(data['frequency'][is_Hnalpha][isnan],
+            3.*data['rms'][is_Hnalpha][isnan]/data['cont'][is_Hnalpha][isnan],
+            linestyle='none',marker=r'$\downarrow$',markersize=20,color='k')
+    if len(data['line2cont'][is_Hnalpha][~isnan]) > 4:
+        # fit line 
+        fit,cov = np.polyfit(data['frequency'][is_Hnalpha][~isnan],
+                             data['line2cont'][is_Hnalpha][~isnan],1,
+                             w=1./data['e_line2cont'][is_Hnalpha][~isnan],cov=True)
+        ax.plot(xfit,np.poly1d(fit)(xfit),'k-',zorder=10,
+                label=r'$F_{{\nu,\rm L}}/F_{{\nu,\rm C}} = ({0:.3f}\pm{1:.3f})\times10^{{-3}}\nu + ({2:.3f}\pm{3:.3f})$'.format(fit[0]*1.e3,np.sqrt(cov[0,0])*1.e3,fit[1],np.sqrt(cov[1,1])))
+        # fit power law
+        fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha][~isnan]),
+                             np.log10(data['line2cont'][is_Hnalpha][~isnan]),1,
+                             w=data['line2cont'][is_Hnalpha][~isnan]*np.log(10.)/data['e_line2cont'][is_Hnalpha][~isnan],cov=True)
+        yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
+        ax.plot(xfit,yfit(xfit),'k--',zorder=10,
+                label=r'$F_{{\nu,\rm L}}/F_{{\nu,\rm C}} \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
+        ax.legend(loc='upper left',fontsize=10)
     # plot stacked
     xlim = ax.get_xlim()
     ax.plot(xlim,
@@ -99,22 +128,31 @@ def main(specfile,sed_plot='cont_sed.pdf',
     fig.tight_layout()
     fig.savefig(line2cont_plot)
     plt.close(fig)
+    print("Done!")
     #
     # Plot Te vs frequency
     #
+    print("Plotting electron temperature...")
     fig, ax = plt.subplots()
-    ax.errorbar(data['frequency'][is_Hnalpha],data['elec_temp'][is_Hnalpha],
-                yerr=data['e_elec_temp'][is_Hnalpha],fmt='o',color='k')
-    # fit line 
-    fit,cov = np.polyfit(data['frequency'][is_Hnalpha],data['elec_temp'][is_Hnalpha],1,w=1./data['e_elec_temp'][is_Hnalpha],cov=True)
-    ax.plot(data['frequency'][is_Hnalpha],np.poly1d(fit)(data['frequency'][is_Hnalpha]),'k-',zorder=10,
-            label=r'$T_e = ({0:.2f}\pm{1:.2f})\nu + ({2:.1f}\pm{3:.1f})$'.format(fit[0],np.sqrt(cov[0,0]),fit[1],np.sqrt(cov[1,1])))
-    # fit power law
-    fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha]),np.log10(data['elec_temp'][is_Hnalpha]),1,w=data['elec_temp'][is_Hnalpha]*np.log(10.)/data['e_elec_temp'][is_Hnalpha],cov=True)
-    yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
-    ax.plot(data['frequency'][is_Hnalpha],yfit(data['frequency'][is_Hnalpha]),'k--',zorder=10,
-            label=r'$T_e \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
-    ax.legend(loc='upper left',fontsize=10)
+    plt.xlim(xmin,xmax)
+    isnan = np.isnan(data['elec_temp'][is_Hnalpha])
+    ax.errorbar(data['frequency'][is_Hnalpha][~isnan],data['elec_temp'][is_Hnalpha][~isnan],
+                yerr=data['e_elec_temp'][is_Hnalpha][~isnan],fmt='o',color='k')
+    if len(data['elec_temp'][is_Hnalpha][~isnan]) > 4:
+        # fit line 
+        fit,cov = np.polyfit(data['frequency'][is_Hnalpha][~isnan],
+                             data['elec_temp'][is_Hnalpha][~isnan],1,
+                             w=1./data['e_elec_temp'][is_Hnalpha][~isnan],cov=True)
+        ax.plot(xfit,np.poly1d(fit)(xfit),'k-',zorder=10,
+                label=r'$T_e = ({0:.2f}\pm{1:.2f})\nu + ({2:.1f}\pm{3:.1f})$'.format(fit[0],np.sqrt(cov[0,0]),fit[1],np.sqrt(cov[1,1])))
+        # fit power law
+        fit,cov = np.polyfit(np.log10(data['frequency'][is_Hnalpha][~isnan]),
+                             np.log10(data['elec_temp'][is_Hnalpha][~isnan]),1,
+                             w=data['elec_temp'][is_Hnalpha][~isnan]*np.log(10.)/data['e_elec_temp'][is_Hnalpha][~isnan],cov=True)
+        yfit = lambda x: 10.**np.poly1d(fit)(np.log10(x))
+        ax.plot(xfit,yfit(xfit),'k--',zorder=10,
+                label=r'$T_e \propto \nu^{{({0:.2f}\pm{1:.2f})}}$'.format(fit[0],np.sqrt(cov[0,0])))
+        ax.legend(loc='upper left',fontsize=10)
     # plot stacked
     xlim = ax.get_xlim()
     ax.plot(xlim,
@@ -129,4 +167,5 @@ def main(specfile,sed_plot='cont_sed.pdf',
     fig.tight_layout()
     fig.savefig(te_plot)
     plt.close(fig)
+    print("Done!")
     plt.ion()
